Refactor: split Avalonia formatting out of OcrFixLineResult
OcrFixLineResult is now pure data (+ GetText); the Avalonia TextBlock formatter
moved to a UI-side extension method (OcrFixLineResultExtensions.GetFormattedText),
so the result type can move to libuilogic. Caller (OcrSubtitleItem) unchanged via
same-namespace extension.
